Apply suggestions made on article

+++ final/app/src/main/java/com/example/makeitso/model/service/StorageService.kt
@@ -26,7 +26,7 @@ interface StorageService {
   suspend fun update(task: Task)
   suspend fun delete(taskId: String)
   suspend fun getCompletedTasksCount(): Long
-  suspend fun getImportantCompletedTasksCount(): Int
+  suspend fun getImportantCompletedTasksCount(): Long
   suspend fun getAverageCompletionTime(): Long
   suspend fun getMediumHighTasksToCompleteCount(): Long
 }
